Tests fix for 9ec8bab72e
The code in 9ec8bab72e is correct, but tests are not, they were not
actually testing the required part as new tests were using RESTModel
instead of AModel.

In fact, usage of additional model was not necessary, beacsue RESTModel
class is recreated before running each of the tests, which means it can
be safely reopend, thus I returned to using RESTModel in all examples.